Reorganize client and communication tests & test data to client package
