+++ client/pedersen_ec_test.go
@@ -1,0 +1,40 @@
+/*
+ * Copyright 2017 XLAB d.o.o.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+package client
+
+import (
+	"math/big"
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+	"github.com/xlab-si/emmy/crypto/groups"
+)
+
+func testPedersenEC(n *big.Int) error {
+	c, err := NewPedersenECClient(testGrpcClientConn, n, groups.P256)
+	if err != nil {
+		return err
+	}
+	return c.Run()
+}
+
+func TestPedersenEC(t *testing.T) {
+	commitVal := big.NewInt(121212121)
+
+	assert.Nil(t, testPedersenEC(commitVal), "should finish without errors")
+}
