+++ client/pseudonymsys_test.go
@@ -15,7 +15,7 @@
  *
  */
 
-package test
+package client
 
 import (
 	"math/big"
@@ -23,7 +23,6 @@ import (
 	"time"
 
 	"github.com/stretchr/testify/assert"
-	"github.com/xlab-si/emmy/client"
 	"github.com/xlab-si/emmy/config"
 	"github.com/xlab-si/emmy/crypto/zkp/schemes/pseudonymsys"
 	"github.com/xlab-si/emmy/server"
@@ -32,13 +31,13 @@ import (
 // TestPseudonymsys requires a running server (it is started in communication_test.go).
 func TestPseudonymsys(t *testing.T) {
 	group := config.LoadGroup("pseudonymsys")
-	caClient, err := client.NewPseudonymsysCAClient(testGrpcClientConn)
+	caClient, err := NewPseudonymsysCAClient(testGrpcClientConn)
 	if err != nil {
 		t.Errorf("Error when initializing NewPseudonymsysCAClient")
 	}
 
 	// usually the endpoint is different from the one used for CA:
-	c1, err := client.NewPseudonymsysClient(testGrpcClientConn)
+	c1, err := NewPseudonymsysClient(testGrpcClientConn)
 	userSecret := c1.GenerateMasterKey()
 
 	p := group.Exp(group.G, userSecret) // this is user's public key
@@ -76,7 +75,7 @@ func TestPseudonymsys(t *testing.T) {
 
 	// register with org2
 	// create a client to communicate with org2
-	caClient1, err := client.NewPseudonymsysCAClient(testGrpcClientConn)
+	caClient1, err := NewPseudonymsysCAClient(testGrpcClientConn)
 	caCertificate1, err := caClient1.ObtainCertificate(userSecret, masterNym)
 	if err != nil {
 		t.Errorf("Error when registering with CA")
@@ -85,7 +84,7 @@ func TestPseudonymsys(t *testing.T) {
 	// c2 connects to the same server as c1, so what we're really testing here is
 	// using transferCredential to authenticate with the same organization and not
 	// transferring credentials to another organization
-	c2, err := client.NewPseudonymsysClient(testGrpcClientConn)
+	c2, err := NewPseudonymsysClient(testGrpcClientConn)
 	nym2, err := c2.GenerateNym(userSecret, caCertificate1, "testRegKey2")
 	if err != nil {
 		t.Errorf(err.Error())
