+++ client/testdata/cspaillierpubkey.txt
