+++ client/testdata/cspaillierseckey.txt
